Fix new compilation errors related to getChildren() list types
The variables expect List<SomeType> exactly, but getChildren() now
returns List<? extends SomeType>, which is not compatible and so causes
a compile error.

I'm guessing that at some point in the past, getChildren() didn't return
the correct generic type, so the intermediate variables with explicit
generic types were added as a workaround. Now that it has a proper
return type, the intermediate variables are no longer necessary.