Update Tycho to version 4.0.13
This requires adding a manual target-platform-configuration, because of
some changes in Tycho 4.0.11 for cases where the thing being compiled
has a Bundle-RequiredExecutionEnvironment that is older than that of the
latest versions of its dependencies. See:
https://github.com/eclipse-tycho/tycho/blob/tycho-4.0.x/RELEASE_NOTES.md#4011

This happens in our case because Rodin 3.9 (and Eclipse 4.34/2023-12, on
which it's based) require Java 17, but our
Bundle-RequiredExecutionEnvironment is still set to JavaSE-11 for
compatibility with older Rodin versions.