Use mvn verify instead of mvn install
"install" installs the built plugin into the Maven local repository.
This is normally not necessary (only necessary when building another
plugin that needs to depend on a locally built version of the ProB
plugin) and Tycho warns about it ("The following locally built units
have been used to resolve project dependencies").

"verify" is the last phase before "install" and is enough to build the
plugin repository and run unit tests.